feat: implement get transactions by product, id, user history and add delete transactions

+++ app/services/transaction_service.py
@@ -39,9 +39,14 @@ def create_exit_transaction(data, user_email, user_id):
     product_repo = ProductsRepository()
     transaction_repo = TransactionsRepository()
 
-    product = product_repo.remove_stock(product_id, quantity)
+    product = product_repo.select_product_by_id(product_id)
     if not product:
-        raise ValueError("Product not found or insufficient stock")
+        raise ValueError("Product not found")
+
+    if product.current_stock < data["quantity"]:
+        raise ValueError("Insufficient stock for transaction")
+
+    product_repo.remove_stock(product_id, quantity)
 
     transaction_hash = generate_transaction_hash(
         product_id=product_id,
@@ -63,3 +68,22 @@ def create_exit_transaction(data, user_email, user_id):
 def get_all_transactions_service():
     repo = TransactionsRepository()
     return repo.select_all_transactions()
+
+def get_transactions_by_product(product_id: int):
+    transaction_repo = TransactionsRepository()
+    return transaction_repo.select_transactions_by_product(product_id)
+
+def delete_transaction_by_id(id: int) -> bool:
+    transaction_repo = TransactionsRepository()
+    return transaction_repo.delete_transaction(id)
+
+def get_transaction_by_id(transaction_id: int):
+    transaction_repo = TransactionsRepository()
+    transaction = transaction_repo.select_transaction_by_id(transaction_id)
+    if not transaction:
+        raise ValueError("Transaction not found")
+    return transaction
+
+def get_transactions_by_user(user_id):
+    repo = TransactionsRepository()
+    return repo.select_transactions_by_user(user_id)
